+++ client/src/app/components/rdio-scanner/admin/admin.service.ts
@@ -21,7 +21,7 @@ import { HttpClient, HttpErrorResponse, HttpHeaders } from '@angular/common/http
 import { EventEmitter, Injectable, OnDestroy } from '@angular/core';
 import { AbstractControl, FormBuilder, FormGroup, ValidationErrors, ValidatorFn, Validators } from '@angular/forms';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { firstValueFrom } from 'rxjs';
+import { firstValueFrom, timer } from 'rxjs';
 import { AppUpdateService } from '../../../shared/update/update.service';
 
 export interface Access {
@@ -539,7 +539,7 @@ export class RdioScannerAdminService implements OnDestroy {
 
                 this.event.emit({ authenticated: this.authenticated });
             } else {
-                setTimeout(() => this.configWebSocketReconnect(), 2000);
+                timer(2000).subscribe(() => this.configWebSocketReconnect());
             }
         };